requirements.txt
streamlit
ipykernel
nltk
scikit-learn
python-dotenv
pandas
numpy

-e .